+++ packages/bloc_error_control/lib/bloc_error_control.dart
@@ -6,8 +6,11 @@ export 'package:meta/meta.dart';
 
 export 'src/annotations/annotations.dart';
 export 'src/exceptions/bloc_canceled_exception.dart';
+export 'src/extensions/cancelable_delay.dart';
+export 'src/interfaces/i_bloc_control.dart';
 export 'src/interfaces/i_cancel_token.dart';
 export 'src/mixins/bloc_error_control_mixin.dart';
 export 'src/models/cancel_request_reasons.dart';
 export 'src/models/event_cancel_token.dart';
 export 'src/models/event_timeout_error.dart';
+export 'src/widgets/bloc_signal_listener.dart';